requirements.txt
python-binance==1.0.19
numpy==1.26.4
requests==2.31.0
urllib3==2.1.0
python-dateutil==2.8.2
pytz==2023.3
